2170B

+++ B_Addition_on_a_Segment.cpp
@@ -1,23 +1,23 @@
 #include <bits/stdc++.h>
 using namespace std;
 
-int main(){
+ int main(){
     ios::sync_with_stdio(false);
     cin.tie(NULL);
 
     int t;
     cin>>t;
     while(t--){
-        int n;
+        long long n;
         cin>>n;
-        vector<int>a(n);
-        int sum=0,cnt=0;
-        for(int i=0;i<n;i++){
+        vector<long long>a(n);
+        long long sum=0,cnt=0;
+        for(long long i=0;i<n;i++){
             cin>>a[i];
             sum+=a[i];
             if(a[i]>0)cnt++;
         }
-        int extra=sum-cnt;
+        long long extra=sum-cnt;
         long long  sub=n-1-extra;
         cout<<cnt-max(0ll,sub)<<"\n";
     }
